Add /models and /tools slash commands
