+++ src/vs/workbench/contrib/chat/browser/actions/chatToolActions.ts
@@ -114,7 +114,7 @@ class SkipToolConfirmation extends ToolConfirmationAction {
 	}
 }
 
-class ConfigureToolsAction extends Action2 {
+export class ConfigureToolsAction extends Action2 {
 	public static ID = 'workbench.action.chat.configureTools';
 
 	constructor() {
